% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_continuous}
\alias{cmap_continuous}
\title{Create continuous colour-Palette for use with ggplot}
\usage{
cmap_continuous(colourmap_name = "rainbow5")
}
\arguments{
\item{colourmap_name}{The name of one of the colourmaps shown via \code{cmap_names}}
}
\value{
returns a function to convert value=x to a hex-value colour.
}
\description{
Creates the colourmap values for your data, from a specified colourmap.
}
