% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_create}
\alias{cmap_create}
\title{Create Colourmap Values}
\usage{
cmap_create(
  colourmap_name = "rainbow5",
  zvals = 1:1000,
  normalise = T,
  output_mode = 1,
  ...
)
}
\arguments{
\item{colourmap_name}{The name of one of the colourmaps shown via \code{cmap_names}}

\item{zvals}{The data values to be colourised}

\item{normalise}{(logical) Whether or not to normalise the data to (0,1)}

\item{output_mode}{There are 3 output modes to choose from:
1 ==> returns a named list
2 ==> returns a detailed data frame
3 ==> returns a simpler data frame}

\item{...}{...}
}
\value{
returns an object mapping your data to colours as per \code{output_mode}
}
\description{
Creates the colourmap values for your data, from a specified colourmap.
}
