% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_names}
\alias{cmap_names}
\title{List Colourmaps}
\usage{
cmap_names()
}
\arguments{
\item{my_str}{A character vector to be cut}

\item{nc}{Number of characters from the left to be kept}
}
\value{
A character vector of extract strings of length nchar=nc
}
\description{
Lists the colourmaps available via \emph{colouRmap}
}
