% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_show}
\alias{cmap_show}
\title{Show a Colourmap}
\usage{
cmap_show(colourmap_name = "none", ncolours = 1000, ...)
}
\arguments{
\item{colourmap_name}{The name of one of the colourmaps shown via \code{cmap_names}}

\item{ncolours}{Number of times to sample the colourmap to create the colourbar}

\item{...}{Other options to pass to the \code{colouRmaps$define_map} function.}
}
\value{
Prints a ggplot object showing the details of the colourmap.
}
\description{
Show a graphical representation of a particular colourmap.
}
