% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_show_all}
\alias{cmap_show_all}
\title{Show All Colourmaps}
\usage{
cmap_show_all()
}
\value{
Prints a ggplot object showing all the colourmaps.
}
\description{
Displays all available colourmaps.
}
