% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{colouRise}
\alias{colouRise}
\title{Wrapper for ggplot2::continuous_scale}
\usage{
colouRise(
  colourmap_name = "roygbiv2",
  aesthetics = "fill",
  scale_name = "gradient",
  na.value = "grey50",
  guide = "colourbar",
  trans = "identity"
)
}
\arguments{
\item{colourmap_name}{The name of one of the colourmaps shown via \code{cmap_names}}

\item{aesthetics}{\itemize{
\item as per \code{ggplot2::continuous_scale}
}}

\item{scale_name}{\itemize{
\item as per \code{ggplot2::continuous_scale}
}}

\item{na.value}{\itemize{
\item as per \code{ggplot2::continuous_scale}
}}

\item{guide}{\itemize{
\item as per \code{ggplot2::continuous_scale}
}}

\item{trans}{\itemize{
\item as per \code{ggplot2::continuous_scale}
}}
}
\value{
a \code{ggplot2::continuous_scale} with a colouRmap palette
}
\description{
Concise call to colourise a ggplot object with a colouRmap
}
