% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colouRmaps}
\alias{colouRmaps}
\title{Colour Map data}
\format{
An Rdata object containing various coulour lookup tables
and functional colour maps.
}
\usage{
colouRmaps
}
\description{
Various Colour-maps for use with GGPLOT2 and other Graphics Tools.
Makes use of the old Astro Starlink colour maps, as well as some
additional bespoke ones.
}
\details{
Creates a set of useful colourmaps to use when creating graphs.
Some of the colour tables were taken from STARLINK's GAIA distribution,
where they have the GNU copyleft.
Further information on STARLINK and GAIA can be found via
http://star-www.rl.ac.uk/ and http://star-www.dur.ac.uk/~pdraper/
}
\keyword{datasets}
