% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colouRmaps.R
\name{cmap_names}
\alias{cmap_names}
\title{List Colourmaps}
\usage{
cmap_names()
}
\value{
A character vector of extract strings of length nchar=nc
}
\description{
Lists the colourmaps available via \emph{colouRmap}
}
